jwt authorization

+++ routes/posts.js
@@ -1,6 +1,20 @@
 import express from "express";
 const postRouter = express.Router();
 import { Post } from "../models/postsModel.js";
+import jwt from "jsonwebtoken";
+
+function auth(req, res, next) {
+  const authHeader = req.headers.authorization;
+  const token = authHeader.split(" ")[1];
+
+  try {
+    const decodedValue = jwt.verify(token, process.env.SECRET);
+    req.user = decodedValue;
+    next();
+  } catch {
+    res.status(401).json({ error: "Invalid token" });
+  }
+}
 
 postRouter.get("/", async (req, res) => {
   try {
@@ -12,10 +26,18 @@ postRouter.get("/", async (req, res) => {
   }
 });
 
-postRouter.post("/", async (req, res) => {
+postRouter.post("/", auth, async (req, res) => {
   try {
+    const author = {
+      name: req.user.userId,
+      age: 21,
+      pfp: ""
+    };
+    req.body.author = author;
     const result = await Post.create(req.body);
     res.send(result).status(200);
+    console.log(req.body);
+    console.log(req.user);
   } catch (err) {
     res.send(err).status(500);
   }
@@ -25,29 +47,36 @@ postRouter.get("/:id", async (req, res) => {
   try {
     const result = await Post.findById(req.params.id);
     res.send(result).status(200);
-  } catch (erre) {
-    res.send(err).status(500);
-  }
-});
-
-postRouter.delete("/:id", async (req, res) => {
-  try {
-    const result = await Post.findByIdAndDelete(req.params.id);
-    res.send(result).status(200);
   } catch (err) {
     res.send(err).status(500);
   }
 });
 
-postRouter.put("/:id", async (req, res) => {
+postRouter.delete("/:id", auth, async (req, res) => {
   try {
-    const result = await Post.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-    console.log("body", req.body);
-    console.log("Params id", req.params.id);
-    console.log("Result", result);
-    res.send(result).status(200);
+    const post = await Post.findById(req.params.id);
+    if (post.author.name.toString() === req.user.userId) {
+      const result = await Post.findByIdAndDelete(req.params.id);
+      res.send(result).status(200);
+    } else {
+      res.status(401).json({ error: "error, unauthorized" });
+    }
+  } catch (err) {
+    res.send(err).status(500);
+  }
+});
+
+postRouter.put("/:id", auth, async (req, res) => {
+  try {
+    const post = await Post.findById(req.params.id);
+    if (post.author.name.toString() === req.user.userId) {
+      const result = await Post.findByIdAndUpdate(req.params.id, req.body, {
+        new: true
+      });
+      res.send(result).status(200);
+    } else {
+      res.status(401).json({ error: "error, unauthorized" });
+    }
   } catch (err) {
     res.send(err).status(500);
   }
